Refactor HeaderBar and HeaderLogo components
